backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
openai==1.3.8
tweepy==4.14.0
pandas==2.1.4
pydantic==2.5.2
python-multipart==0.0.6